Remove the protocol ResultShowable
Move the method makeResult() into CardDeck
