+++ CardGame/CardGame/Deck/CardDeck.swift
@@ -45,4 +45,20 @@ struct CardDeck: Playable {
         shuffle()
     }
 
+    mutating func makeResult(of menu: MenuItem) -> String {
+        switch menu {
+        case .reset:
+            reset()
+            return "카드 전체를 초기화했습니다."
+        case .shuffle:
+            shuffle()
+            return "전체 \(self.count())장의 카드를 섞었습니다."
+        case .draw:
+            let card = removeOne()
+            return "\(card)\n총 \(self.count())장의 카드가 남아있습니다."
+        case .quit:
+            return "---- END ----"
+        }
+    }
+
 }
